server/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
annotated-types==0.7.0
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.5.0
async-timeout==4.0.3
attrs==21.2.0
awscli==1.22.34
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
botocore==1.23.34
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
docutils==0.17.1
duplicity==0.8.21
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
fasteners==0.14.1
fonttools==4.29.1
frozenlist==1.2.0
fs==2.4.12
future==0.18.2
gast==0.5.2
gpg==1.16.0
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.2.0
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
json5==0.9.14
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libcomps==0.1.15
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multidict==5.1.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.8
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==2.2.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
proton-core==0.1.16
proton-keyring-linux==0.0.2
proton-keyring-linux-secretservice==0.0.2
proton-vpn-api-core==0.27.2
proton-vpn-connection==0.14.4
proton-vpn-gtk-app==4.3.3
proton-vpn-killswitch==0.4.0
proton-vpn-killswitch-network-manager==0.4.5
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.4.2
proton-vpn-network-manager-openvpn==0.0.7
proton-vpn-session==0.6.7
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyarrow==15.0.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-multipart==0.0.9
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
redis==5.0.8
reportlab==3.6.8
requests==2.25.1
rich==13.7.1
roman==3.3
rpm==4.17.0
rq==1.16.2
rsa==4.8
s3transfer==0.5.0
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
sentry-sdk==1.4.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
starlette==0.37.2
sympy==1.9
systemd-python==234
tabulate==0.8.9
terminado==0.13.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.12.2
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
uvloop==0.19.0
wadllib==1.3.6
watchgod==0.8.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
widgetsnbextension==2.0.0
wsproto==1.0.0
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
